+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.6.5
 * Bugfix
   - fix: `StkTableColumn['width']` not work correctly.
+  - fix: when props.showHeaderOverflow not set, the header row height not work correctly in virtual mode .
 
 ## 0.6.4
 * Feature